+++ sismic/testing/tester.py
@@ -24,9 +24,8 @@ def _teststory_from_macrostep(macrostep: MacroStep) -> Story:
                                event=macrostep.event))
         for state in microstep.entered_states:
             story.append(Event('state entered', state=state))
-
-    for event in macrostep.sent_events:
-        story.append(Event('event sent', event=event))
+        for event in microstep.sent_events:
+            story.append(Event('event sent', event=event))
 
     story.append(Event('step ended'))
     return story